+++ src/components/common/detail/Buttons/Buttons.tsx
